requirements.txt
# Core API
fastapi
uvicorn
python-multipart
requests

# Image Processing
opencv-python
numpy
Pillow
PyMuPDF

# OCR & ML Models
paddlepaddle
paddleocr
transformers>=4.30.0
torch==2.6.0

# Utilities
aiofiles
